feat: move network utilities in `lib` to separate folder, update imports
